+++ Mods ChoGGi/Gagarin Logos/metadata.lua
@@ -1,0 +1,15 @@
+return PlaceObj("ModDef", {
+	"title", "Gagarin Logos v0.1",
+	"version", 1,
+	"saved", 1542542400,
+	"image", "Preview.png",
+	"id", "ChoGGi_GagarinLogos",
+	"steam_id", "1578000203",
+	"author", "ChoGGi",
+	"lua_revision", LuaRevision,
+	"code", {
+		"Code/Script.lua",
+	},
+	"description", [[Someone forgot to include these in base game (only included the logo defs in the Space race DLC).
+They'll be unlocked next update, but till then.]],
+})